+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## ??? (Not Released Yet)
 
 * Recherche des contacts: on ignore les contacts supprimés. Cela n'est pas censé arriver, mais il semblerait que ce soit le cas sur certains environnements.
+* Script fix_contribution_date.php pour corriger les dates des contributions.
 
 ## 1.6.0